requirements.txt

fastapi
uvicorn[standard]
pydantic
python-multipart
pdfplumber
python-docx
docx2txt
pytesseract
Pillow
spacy

sentence-transformers
transformers
torch
scikit-learn
pandas
numpy

aiohttp
requests
python-dotenv
sqlalchemy
databases
alembic
redis
celery
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-jose
passlib
psycopg2-binary
